orchestrator-service/requirements.txt
pylint
pytest
sqlalchemy
psycopg2-binary==2.9.6
Django==4.2
djangorestframework>=3.12.0
drf-yasg
django-cors-headers
django-prometheus
structlog
redis>=4.0.0
django-redis>=5.0.0
requests>=2.28.0
